refactor(05-01): split layout.rs into src/layout/ directory module
- Extract LayoutRenderer to layout/template.rs
- Extract VariableBuilder to layout/variables.rs
- Extract presets to layout/presets.rs
- Extract color/formatting helpers to layout/format.rs
- Relocate all tests to layout/tests.rs
- Add pub use re-exports in mod.rs preserving public API
- All 308 existing tests pass unchanged
